Added the creators for the other tables of the database, moved the Creator to the directory Test

+++ app/src/main/java/com/github/ricardopolit/sipmanager/data/asset/Asset.kt
@@ -24,7 +24,7 @@ data class Asset(
         val color: String,
 
         @ColumnInfo(name = "income_type")
-        val incomeType: Long,
+        val incomeType: Int,
 
         @ColumnInfo(name = "active")
         val active: Boolean = true
